[BE] refactor: 패키지 `mapper`, `validator` 분리 (#598)
* refactor: 패키지 `mapper`, `validator` 분리

* refactor: TemplateMapper 패키지명 변경

* refactor: PreviewGenerator 사용되는 패키지로 이동
