+++ backend/src/main/java/reviewme/review/service/ReviewDetailLookupService.java
@@ -9,7 +9,7 @@ import reviewme.review.repository.ReviewRepository;
 import reviewme.review.service.dto.response.detail.ReviewDetailResponse;
 import reviewme.review.service.exception.ReviewGroupUnauthorizedException;
 import reviewme.review.service.exception.ReviewNotFoundByIdAndGroupException;
-import reviewme.review.service.module.ReviewDetailMapper;
+import reviewme.review.service.mapper.ReviewDetailMapper;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
