+++ backend/src/main/java/reviewme/review/service/ReviewListLookupService.java
@@ -8,7 +8,7 @@ import reviewme.review.service.exception.ReviewGroupNotFoundByReviewRequestCodeE
 import reviewme.review.service.dto.response.list.ReceivedReviewsResponse;
 import reviewme.review.service.dto.response.list.ReviewListElementResponse;
 import reviewme.review.service.exception.ReviewGroupUnauthorizedException;
-import reviewme.review.service.module.ReviewListMapper;
+import reviewme.review.service.mapper.ReviewListMapper;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
