+++ backend/src/test/java/reviewme/review/service/ReviewPreviewGeneratorTest.java
@@ -1,4 +1,4 @@
-package reviewme.review.service.module;
+package reviewme.review.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import reviewme.review.domain.TextAnswer;
+import reviewme.review.service.mapper.ReviewPreviewGenerator;
 
 class ReviewPreviewGeneratorTest {
